requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.6.20
Django==3.0.3
django-bootstrap4==2.2.0
importlib-metadata==1.7.0
pytube3==9.6.4
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
typing-extensions==3.7.4.2
zipp==3.1.0
